chore: finishes metadata migration in cloud folder
