+++ src/app/cloud/[team-slug]/[project-slug]/(tabs)/settings/billing/page.tsx
@@ -1,11 +1,19 @@
 import { Metadata } from 'next'
 
+import { mergeOpenGraph } from '@root/seo/mergeOpenGraph'
 import { ProjectBillingPage } from './client_page'
 
 export default props => {
   return <ProjectBillingPage {...props} />
 }
 
-export const metadata: Metadata = {
-  title: 'Billing',
+export async function generateMetadata({
+  params: { 'team-slug': teamSlug, 'project-slug': projectSlug },
+}) {
+  return {
+    title: 'Billing',
+    openGraph: mergeOpenGraph({
+      url: `/cloud/${teamSlug}/${projectSlug}/settings/billing`,
+    }),
+  }
 }
