+++ src/app/cloud/[team-slug]/[project-slug]/(tabs)/settings/domains/page.tsx
@@ -1,11 +1,19 @@
 import { Metadata } from 'next'
 
+import { mergeOpenGraph } from '@root/seo/mergeOpenGraph'
 import { ProjectDomainsPage } from './client_page'
 
 export default props => {
   return <ProjectDomainsPage {...props} />
 }
 
-export const metadata: Metadata = {
-  title: 'Domains',
+export async function generateMetadata({
+  params: { 'team-slug': teamSlug, 'project-slug': projectSlug },
+}) {
+  return {
+    title: 'Domains',
+    openGraph: mergeOpenGraph({
+      url: `/cloud/${teamSlug}/${projectSlug}/settings/domains`,
+    }),
+  }
 }
