+++ src/app/cloud/[team-slug]/[project-slug]/(tabs)/settings/environment-variables/page.tsx
@@ -1,11 +1,19 @@
 import { Metadata } from 'next'
 
+import { mergeOpenGraph } from '@root/seo/mergeOpenGraph'
 import { ProjectEnvPage } from './client_page'
 
 export default props => {
   return <ProjectEnvPage {...props} />
 }
 
-export const metadata: Metadata = {
-  title: 'Environment Variables',
+export async function generateMetadata({
+  params: { 'team-slug': teamSlug, 'project-slug': projectSlug },
+}) {
+  return {
+    title: 'Environment Variables',
+    openGraph: mergeOpenGraph({
+      url: `/cloud/${teamSlug}/${projectSlug}/settings/environment-variables`,
+    }),
+  }
 }
