+++ src/app/cloud/[team-slug]/[project-slug]/(tabs)/settings/ownership/page.tsx
@@ -1,11 +1,19 @@
 import { Metadata } from 'next'
 
+import { mergeOpenGraph } from '@root/seo/mergeOpenGraph'
 import { ProjectOwnershipPage } from './client_page'
 
 export default props => {
   return <ProjectOwnershipPage {...props} />
 }
 
-export const metadata: Metadata = {
-  title: 'Ownership',
+export async function generateMetadata({
+  params: { 'team-slug': teamSlug, 'project-slug': projectSlug },
+}) {
+  return {
+    title: 'Ownership',
+    openGraph: mergeOpenGraph({
+      url: `/cloud/${teamSlug}/${projectSlug}/settings/ownership`,
+    }),
+  }
 }
