+++ src/app/cloud/[team-slug]/settings/page.tsx
@@ -1,13 +1,17 @@
 import { Metadata } from 'next'
 
+import { mergeOpenGraph } from '@root/seo/mergeOpenGraph'
 import { TeamSettingsPage } from './client_page'
 
 export default props => {
   return <TeamSettingsPage {...props} />
 }
 
-export async function generateMetadata({ params: { 'team-slug': slug } }): Promise<Metadata> {
+export async function generateMetadata({ params: { 'team-slug': teamSlug } }): Promise<Metadata> {
   return {
-    title: `${slug} - Team Settings`,
+    title: `${teamSlug} - Team Settings`,
+    openGraph: mergeOpenGraph({
+      url: `/cloud/${teamSlug}/settings`,
+    }),
   }
 }
