+++ src/app/cloud/page.tsx
@@ -1,5 +1,6 @@
 import { Metadata } from 'next'
 
+import { mergeOpenGraph } from '@root/seo/mergeOpenGraph'
 import { CloudHomePage } from './client_page'
 
 export default props => {
@@ -8,4 +9,7 @@ export default props => {
 
 export const metadata: Metadata = {
   title: 'Home | Payload Cloud',
+  openGraph: mergeOpenGraph({
+    url: '/cloud',
+  }),
 }
