+++ src/app/cloud/settings/page.tsx
@@ -1,11 +1,17 @@
 import { Metadata } from 'next'
 
+import { mergeOpenGraph } from '@root/seo/mergeOpenGraph'
 import { SettingsPage } from './client_page'
 
 export default props => {
   return <SettingsPage {...props} />
 }
 
-export const metadata: Metadata = {
-  title: 'My Account',
+export async function generateMetadata({ params: { 'team-slug': teamSlug } }): Promise<Metadata> {
+  return {
+    title: `${teamSlug} - My Account`,
+    openGraph: mergeOpenGraph({
+      url: `/cloud/settings`,
+    }),
+  }
 }
