+++ src/app/new/clone/page.tsx
@@ -2,6 +2,8 @@ import React from 'react'
 import { NewProjectBlock } from '@blocks/NewProject'
 import { Metadata } from 'next'
 
+import { mergeOpenGraph } from '@root/seo/mergeOpenGraph'
+
 const ProjectFromTemplate: React.FC = () => {
   return <NewProjectBlock />
 }
@@ -10,4 +12,7 @@ export default ProjectFromTemplate
 
 export const metadata: Metadata = {
   title: 'Clone Template | Payload Cloud',
+  openGraph: mergeOpenGraph({
+    url: '/new/clone',
+  }),
 }
