+++ src/app/new/page.tsx
@@ -2,6 +2,8 @@ import React from 'react'
 import { NewProjectBlock } from '@blocks/NewProject'
 import { Metadata } from 'next'
 
+import { mergeOpenGraph } from '@root/seo/mergeOpenGraph'
+
 const NewProjectPage: React.FC = () => {
   return <NewProjectBlock headingElement="h1" />
 }
@@ -10,4 +12,7 @@ export default NewProjectPage
 
 export const metadata: Metadata = {
   title: 'New Project | Payload Cloud',
+  openGraph: mergeOpenGraph({
+    url: '/new',
+  }),
 }
