--- src/app/test/page.tsx
@@ -1,5 +1,0 @@
-'use client'
-
-export default () => {
-  return <h1>test page</h1>
-}
